rename helper-files (utils, config) and move corresponding to store and router folders
