+++ src/store/auth/actions.js
@@ -1,7 +1,7 @@
 import { setAccessToken, deleteAccessToken } from '@/settings/axiosSettings'; // eslint-disable-line import/no-cycle
 import { router } from '@/router'; // eslint-disable-line import/no-cycle
 import { addQueryItems } from '@/router/utils'; // eslint-disable-line import/no-cycle
-import { dispatchFromRoot, commitFromRoot } from '@/store/store-utils'; // eslint-disable-line import/no-cycle
+import { dispatchFromRoot, commitFromRoot } from '@/store/utils'; // eslint-disable-line import/no-cycle
 
 export default {
   async _signUp(user) {
