+++ src/store/config.js
@@ -1,0 +1,1 @@
+export const MIN_SEARCH_SYMBOLS = 3;
