+++ src/utils/misc.js
@@ -20,11 +20,3 @@ export class DeferredBoolean {
       .catch(() => false);
   }
 }
-
-export function getDefaultValue(optionsObj, option) {
-  return optionsObj[option].default;
-}
-
-export function checkDefaultValue(optionsObj, option, value) {
-  return value === getDefaultValue(optionsObj, option);
-}
